refactoring to aplusstory.fixme package
