+++ app/build.gradle
@@ -30,4 +30,7 @@ dependencies {
     implementation 'com.prolificinteractive:material-calendarview:1.4.3'
     implementation 'com.android.support:preference-v7:28.0.0-alpha1'
     implementation 'com.github.PhilJay:MpAndroidChart:v3.0.2'
+    implementation 'com.squareup:android-times-square:1.6.5@aar'
+    implementation 'com.github.ik024:calendar-lib:2.0.2'
+    implementation 'io.github.memfis19:cadar:0.4.8'
 }
