--- app/src/main/java/com/example/nayunpark/fixme_ui/AlarmPeriodActivity.java
@@ -1,39 +1,0 @@
-package com.example.nayunpark.fixme_ui;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.Spinner;
-import android.widget.TextView;
-import android.support.v7.widget.Toolbar;
-
-public class AlarmPeriodActivity extends AppCompatActivity {
-
-    Spinner spinner;
-    Toolbar toolbar;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_settings_alarmperiod);
-
-        toolbar = (Toolbar) findViewById(R.id.periodBar);
-        setSupportActionBar(toolbar);
-
-        getSupportActionBar().setDisplayShowTitleEnabled(false);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_full_menu);
-
-        spinner = (Spinner) findViewById(R.id.periodSpinner);
-
-
-        final String[] item = {"선택하세요", "5분마다", "10분마다", "30분마다", "1시간마다", "3시간마다"};
-
-        ArrayAdapter<String> arrayAdapter = new ArrayAdapter<String>(this, android.R.layout.simple_spinner_dropdown_item,item);
-        arrayAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-
-        spinner.setAdapter(arrayAdapter);
-    }
-}
