--- app/src/main/java/com/example/nayunpark/fixme_ui/ChartInfo.java
@@ -1,12 +1,0 @@
-package com.example.nayunpark.fixme_ui;
-
-public class ChartInfo {
-
-    public String percentData, locationName, timeData;
-
-    public ChartInfo(String percentData, String locationName, String timeData) {
-        this.percentData = percentData;
-        this.locationName = locationName;
-        this.timeData = timeData;
-    }
-}
