--- app/src/main/java/com/example/nayunpark/fixme_ui/FootDataSettingsActivity.java
@@ -1,35 +1,0 @@
-package com.example.nayunpark.fixme_ui;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.widget.ArrayAdapter;
-import android.widget.Spinner;
-
-public class FootDataSettingsActivity extends AppCompatActivity {
-
-    Toolbar toolbar;
-    Spinner spinner;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_settings_data);
-
-        toolbar = (Toolbar) findViewById(R.id.toolbar2);
-        setSupportActionBar(toolbar);
-
-        getSupportActionBar().setDisplayShowTitleEnabled(false);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_full_menu);
-
-        spinner = (Spinner) findViewById(R.id.maintainSpinner);
-
-        final String[] item = {"선택하세요", "없음", "1주일", "1개월", "6개월", "1년"};
-
-        ArrayAdapter<String> arrayAdapter = new ArrayAdapter<String>(this, android.R.layout.simple_spinner_dropdown_item,item);
-        arrayAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-
-        spinner.setAdapter(arrayAdapter);
-    }
-}
