--- app/src/main/java/com/example/nayunpark/fixme_ui/MainActivity.java
@@ -1,43 +1,0 @@
-package com.example.nayunpark.fixme_ui;
-
-import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        Button scheduleB = (Button) findViewById(R.id.ScheduleButton);
-        scheduleB.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), ScheduleActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        Button todayFootprintButton = (Button) findViewById(R.id.FootprintButton);
-        todayFootprintButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getApplicationContext(), FootprintActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        Button settingsB = (Button) findViewById(R.id.SettingsButton);
-        settingsB.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), SettingsActivity.class);
-                startActivity(intent);
-            }
-        });
-    }
-}
