--- app/src/main/java/com/example/nayunpark/fixme_ui/ScheduleAlarmActivity.java
@@ -1,39 +1,0 @@
-package com.example.nayunpark.fixme_ui;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.Spinner;
-import android.widget.TextView;
-import android.support.v7.widget.Toolbar;
-
-public class ScheduleAlarmActivity extends AppCompatActivity {
-
-    Spinner spinner;
-    Toolbar toolbar;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_schedule_alarm);
-
-        toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        getSupportActionBar().setDisplayShowTitleEnabled(false);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        getSupportActionBar().setHomeAsUpIndicator(R.drawable.ic_full_menu);
-
-        spinner = (Spinner) findViewById(R.id.alarmSpinner);
-
-
-        final String[] item = {"선택하세요", "1분 전", "5분 전", "10분 전", "30분 전", "1시간 전", "6시간 전"};
-
-        ArrayAdapter<String> arrayAdapter = new ArrayAdapter<String>(this, android.R.layout.simple_spinner_dropdown_item,item);
-        arrayAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-
-        spinner.setAdapter(arrayAdapter);
-    }
-}
