--- app/src/main/java/com/example/nayunpark/fixme_ui/ScheduleFragment.java
@@ -1,207 +1,0 @@
-package com.example.nayunpark.fixme_ui;
-
-import android.app.Activity;
-import android.content.Context;
-import android.content.Intent;
-import android.net.Uri;
-import android.os.Bundle;
-import android.os.Parcel;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.ImageButton;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import java.text.SimpleDateFormat;
-import java.util.Date;
-
-import static android.app.Activity.RESULT_OK;
-
-
-/**
- * A simple {@link Fragment} subclass.
- * Activities that contain this fragment must implement the
- * {@link ScheduleFragment.OnFragmentInteractionListener} interface
- * to handle interaction events.
- * Use the {@link ScheduleFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
-public class ScheduleFragment extends Fragment {
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
-
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
-
-    Date today = new Date();
-    SimpleDateFormat date = new SimpleDateFormat("yyyy.MM.dd");
-    String repeatDate = "none";
-    int repeatState = 0;
-    private int REQUEST_RESULT = 1;
-
-    private OnFragmentInteractionListener mListener;
-
-    public ScheduleFragment() {
-        // Required empty public constructor
-    }
-
-
-    // TODO: Rename and change types and number of parameters
-    public static ScheduleFragment newInstance() {
-        return new ScheduleFragment();
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        Log.d(ScheduleFragment.class.toString(),ScheduleFragment.class.toString() + " created");
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
-
-
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        View returnView = inflater.inflate(R.layout.fragment_schedule, container, false);
-
-        TextView textView = (TextView) returnView.findViewById(R.id.scheduleDate);
-        textView.setText(date.format(today));
-
-        TextView repeationDetail = (TextView) returnView.findViewById(R.id.repeationDetail);
-        repeationDetail.setText(repeatDate);
-
-        ImageButton timeButton = (ImageButton) returnView.findViewById(R.id.timeButton);
-        timeButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(getActivity(), ScheduleTimeActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        ImageButton alarmButton = (ImageButton) returnView.findViewById(R.id.alarmButton);
-        alarmButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getActivity(), ScheduleAlarmActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        ImageButton colorButton = (ImageButton) returnView.findViewById(R.id.colorButton);
-        colorButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getActivity(), ScheduleColorActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        ImageButton repeatButton = (ImageButton) returnView.findViewById(R.id.repeationButton);
-        repeatButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getActivity(), ScheduleRepeationActivity.class);
-                startActivityForResult(intent, REQUEST_RESULT);
-            }
-        });
-
-        Button deleteButton = (Button) returnView.findViewById(R.id.deleteButton);
-        deleteButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                FragmentManager fragmentManager = getActivity().getSupportFragmentManager();
-                fragmentManager.beginTransaction().hide(ScheduleFragment.this).commit();
-                fragmentManager.popBackStack();
-            }
-        });
-
-
-        return returnView;
-    }
-
-    // TODO: Rename method, update argument and hook method into UI event
-    public void onButtonPressed(Uri uri) {
-        if (mListener != null) {
-            mListener.onFragmentInteraction(uri);
-        }
-    }
-
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-        if(requestCode == REQUEST_RESULT) {
-            if(resultCode == RESULT_OK) {
-                Bundle bundle = this.getArguments();
-                if(bundle != null) {
-                    ScheduleData scheduleData = bundle.getParcelable("schedule");
-
-                    repeatDate = scheduleData.getRepeatDate();
-                    repeatState = scheduleData.getRepeatState();
-                }
-                refresh();
-                Toast.makeText(getContext(), "repeatDate: "+repeatDate+", repeatState: "+repeatState, Toast.LENGTH_SHORT).show();
-            }
-        }
-    }
-
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        if (context instanceof OnFragmentInteractionListener) {
-            mListener = (OnFragmentInteractionListener) context;
-        } else {
-            /*throw new RuntimeException(context.toString()
-                    + " must implement OnFragmentInteractionListener");*/
-        }
-    }
-
-
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        mListener = null;
-    }
-
-
-
-
-    /**
-     * This interface must be implemented by activities that contain this
-     * fragment to allow an interaction in this fragment to be communicated
-     * to the activity and potentially other fragments contained in that
-     * activity.
-     * <p>
-     * See the Android Training lesson <a href=
-     * "http://developer.android.com/training/basics/fragments/communicating.html"
-     * >Communicating with Other Fragments</a> for more information.
-     */
-    public interface OnFragmentInteractionListener {
-        // TODO: Update argument type and name
-        void onFragmentInteraction(Uri uri);
-    }
-
-    private void refresh() {
-        FragmentTransaction transaction = getFragmentManager().beginTransaction();
-        transaction.detach(this).attach(this).commit();
-    }
-
-
-}
