--- app/src/main/java/com/example/nayunpark/fixme_ui/SettingsColorInfo.java
@@ -1,12 +1,0 @@
-package com.example.nayunpark.fixme_ui;
-
-public class SettingsColorInfo {
-
-    public int colorId;
-    public String colorName;
-
-    public SettingsColorInfo(int colorId, String colorName) {
-        this.colorId = colorId;
-        this.colorName = colorName;
-    }
-}
