--- app/src/main/java/com/example/nayunpark/fixme_ui/SplashActivity.java
@@ -1,21 +1,0 @@
-package com.example.nayunpark.fixme_ui;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-
-public class SplashActivity extends Activity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        try {
-            Thread.sleep(2000);
-        } catch(InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        startActivity(new Intent(this, MainActivity.class));
-        finish();
-    }
-}
