+++ app/src/main/res/layout/activity_chart.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".FootprintActivity">
+    tools:context="com.aplusstory.fixme.FootprintActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
