+++ app/src/main/res/layout/activity_favorite.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".FavoriteActivity">
+    tools:context="com.aplusstory.fixme.FavoriteActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
