+++ app/src/main/res/layout/activity_footprint_routine.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".FootprintRoutineActivity">
+    tools:context="com.aplusstory.fixme.FootprintRoutineActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
