+++ app/src/main/res/layout/fragment_blank.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BlankFragment"
+    tools:context="com.aplusstory.fixme.BlankFragment"
     android:id="@+id/fragment_blank"
     android:background="@android:color/white">
