+++ app/src/main/res/layout/fragment_schedule.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/fragment_schedule"
-    tools:context=".ScheduleFragment"
+    tools:context="com.aplusstory.fixme.ScheduleFragment"
     android:background="@android:color/white">
