+++ app/src/main/res/layout/fragment_schedule_repeat_weekly.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScheduleRepeatWeeklyFragment">
+    tools:context="com.aplusstory.fixme.ScheduleRepeatWeeklyFragment">
 
     <!-- TODO: Update blank fragment layout -->
